+++ backend/src/main/java/com/debede/backend/repository/BoletaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoletaRepository extends JpaRepository<Boleta, Long> {
+public interface BoletaRepository extends JpaRepository<Boleta, Integer> {
 
 }
